Handle timestamps/durations as u32, improve errors, detect out-of-range dates

+++ src/ast.rs
@@ -86,7 +86,7 @@ impl_from_variant!(ArrayAccess, Expr);
 // Duration (relative block height or time)
 #[derive(Debug, Clone)]
 pub enum Duration {
-    BlockHeight(usize),
+    BlockHeight(u32),
     BlockTime {
         parts: Vec<DurationPart>,
         heightwise: bool,
